Restructure all prices to have associated units
Witnesses now have their own type of unit.
The basic fee is now a unit of type DAY rather than fixed, but then
has a new fixed_field associated with it (and a 0 fee_per_unit). This
allows a more transparent allocation of total days spent on the case
between the different levels of basic fee.
Retrieving fee types now also returns a list of applicable units,
which can then be entered separately to calculate the cost in each
case (or that is how it will work).