Allow filtering of fee types by the scheme, scenario, offence, advocate type
The idea of this is to allow the client to retrieve a list of relevant
fees once the original parameters have been selected. The filtering by
these types hase to be done separately rather than via generic filters
as otherwise it creates a join on the prices table for every filter field
and will not require that each price that it is joining on satisfy all
of the field values.