Add fixed_percent modifiers and modifier priority
Modifier priority allows for certain modifiers to be applied after
others. Modifiers with the same priority will be applied to the same
base price - those with higher priority will be applied to the price
plus any lower priority modifiers.

Fixed percent behvaes slightly differently to percent per unit - in
order to be applicable, the count must be within the range, rather than
just greater than the limit_from as in percent per unit.

Added in 50% discount to discontinuance when there is 0 PPE.