Add an explicit strict_range attribute to modifiers
Rather than only applying the modifier if the value is strictly
within the range (rather than if the value "covers" the range) when
it is a fixed_fee, now there is an explicit switch in the form of
strict_range. If strict_range is true, the modifier will only apply
if the value is both greater than limit_from and less than limit_to
(or equal to, both ways).