Change the way in which fee types with multiple units are dealt with
Previously the client would request calculation for each unit
separately, whereas now they request calculation for all applicable
units in the same request. This is to allow for a difference in the
way that the totals for these units may be combined. For instance,
in the AGFS basic fee the totals are summed, whereas for the LGFS
basic fee the maximum of the totals is used.