Remove prices for invalid scenario/feetype combinations
For: appeal against conviction, appeal against sentence, committal for
sentence, contempt, and breach of crown court order, it should not be
possible to claim the associated fee type with anything other than the
associated scenario. It was previously possible to claim them with
any scenario (AGFS), but this has been rectified.

For the scenario of elected case not proceeded, it should only be possible
to claim a basic fee or a fee for confiscation. No other misc fees or fixed
fees should be chargeable. Previously all misc fees and fixed fees were
chargeable under AGFS 10.