Fix interim fees for trial/retrial start
Per day fees for interim fees for trial/retrial start should not
be chargeable for more than one day, and no fees should be chargeable
at all if the trial length is less than 10.

To reflect this, a required modifier of trial_length >= 10 has been
added to all prices for this scenario, and the 'DAY' prices for values
higher than 1 have been deleted.