Clear data on build
was hitting a UNIQUE constraint error because
data already existed. I do not believe this effects
circleCI builds (I assume because they are starting with
a cleared database).

Since this is a buildtime change it should not have any
impact if we explicitly clear the DB prior to "seeding"
the database.