Split test run into 2 containers
There are 9700 tests in total at the moment
taking over 12 minutes. This will increase
significantly when adding a new fee scheme.

How:
add parrellism of 2 and glob the test dir
files for splitting by file name.

However, sed is needed to convert the file
names to python dotted paths as file paths
are not recognised by django `test` runs

Also had to --noinput to ensure any existing
test db is destroyed. I believe this was only
needed because the test db had not been destroyed
and was not gitignored in a previous (since removed)
commit.