Add documentation for adding new scheme
Tools existed before but no documentation.

There is a command line tool copyscheme which
works fine, however there is no documentation
as to what to do with the data it generates in the
DB - which will be overwritten by the fixtures on
build and deploy of the app.

This appears to be because the `price` fixture should then
be generated using a standard django `dumpdata` utility
after the scheme is copied.