Add scheme 12 cracked trial scenario prices
Cracked trial advocate fees for scheme 12
no are the 100% of the equivalent trial sceanrio
fee price regardless of the third in which the
trial cracked.

Note: Ideally we should remove the modifier effect
for this scenario since THIRD_CRACKED no longer
has any impact on the value. However CLAR
is being brought in a very rushed manor and no
one with the tech stack and domain knowledge to
do this quickly is available.