Replace CalculatorSchema with CalcuatorQuerySerializer
CalculatorSchema not viable for move to openAPI3 so
using a custom query serializer for this custom endpoint.

The dynmic creation of parameters is probably not an
ideal solution and does result in docker build time errors
due the DB being created from scratch as part of the build.
This is apparently because the CalculateQuerySerializer iterates
over `Unit` and `ModifierType` objects but their tables have not yet
been created at build time, causing it to raise this error:

```
django.db.utils.OperationalError: no such table: calculator_unit
```

By ignoring the error build passes and the documentation is successfully
generated at run time regardless, presumebly because the tables have been
created by that point.