Remove dynamic generation of serializer fields
Two sets of fields on the CalculatorQuerySerializer are being
dynamically created on the class. This is done because the serializer
dict is immutable and so they cannot be programmatically added when the
serializer has been instantiated.

However the approach used to do this, by calling exec() is unclean.

To improve it, this commit removes the dynamic generation and instead
hardcodes the necessary fields. This is less elegant but safer.