Start of scenario view
The scenario is a single string containing an amalgamation of various pieces
of information (bill type, fee type, etc). Ideally this will be fixed at the
model level.

To get a better view on this information a 'scenario' view is created that will
attempt to split this information into its component parts. This is work in
progress.