Split parallelized tests by timings
CircleCI runs the test suite in parallel across two containers. It
currently splits the tests by the default setting which is inefficient -
it results in one container which takes twice as long to complete as the
other.

This change adds the `unittest-xml-reporting` library in order to
generate xml test reports. These are persisted in CircleCI allowing the
tests to instead be split by timings.

In order to do this the command used by CircleCI to identify which tests
to run needed to be changed as suggested in guidance here:
https://support.circleci.com/hc/en-us/articles/360048786831-Use-test-splitting-with-Python-Django-tests

This results in parallelization which is more efficient and results in
both containers taking approximately the same amount of time, thus
reducing the overall elapsed time of the tests.