Use SimpleTestCase in calculator tests
Django's `SimpleTestCase` is a subclass of `unittest.TestCase` which runs
faster. This commit replaces `TestCase` with `SimpleTestCase` in the
`CalculatorTestCase` class which is inherited from by all other calculator
tests.

By default `SimpleTestCase` does not allow database queries, so this is
enabled by adding the `default` database to the class.

Locally this reduces the time taken to run the whole test suite from
approximately 11 minutes to 5 minutes.