CFP-577 Run Flake8 in CircleCI workflow
Adds a `linting` job to the CircleCI config and runs it as part of the
`build-test-and-approval-deploy` workflow in parallel with the `test`
job.

Due to the version of flake8 used on CircleCI it has been necessary to
tweak the syntax of how we ignore certain code (specifically, methods
with annotations).