CFP-557 Improve caching in CircleCI
The `test` and `linting` jobs both independently install python
dependencies (multiple times in the case of the `test` job as it is
run on two containers.)

To make this more efficient, this commit adds caching to the
`setup_python_env` command. A cached version of the dependencies is
used if one exists, any new dependencies are installed, and the cache is
then saved. This command can be used across multiple jobs to prevent
the repeated installation of dependencies.