CFP-561 Remove redundant live-1 hosts
Post EKS-migration, Fee Calculator deploys to dev, staging and
production with two hosts, one in the format

laa-fee-calculator-dev.apps.live-1.cloud-platform.service.justice.gov.uk

and the other in the format

dev.laa-fee-calculator.service.justice.gov.uk.

The first of these is redundant, and is being flagged by the Cloud
Platform team as an infringement.

This commit removes the config for those hosts from the relevant
ingress file for each environment.