Update deprecated GitHub Action
The GitHub CodeQL Action v1 will be deprecated in December 2022. This
is used in the `docker_image_scan` job which runs nightly to scan our
docker images with Snyk.

To resolve this and avoid loss of service in future, this commit
replaces `github/codeql-action/upload-sarif@v1` with
`github/codeql-action/upload-sarif@v2`.