Test that only one fee scheme is returned
The search `?type=AGFS&case_date=2012-04-02` only returns a single fee scheme
but the test only checks that the first element returned is the correct one. As
it happens a wider search, such as `?type=AGFS` will also return the same
scheme as the first element so this test would still pass if the filter is
broken. It is therefore necessary to ensure that only a single item is
returned.