Use django_filters for schemes
django_filters is used for some views but not all. To isolate the filtering
logic, a new filter is used for SchemeViewSet.

Note, the previously the values of case_date and type were validated and raised
exceptions if they were invalid. This check is now done by django_filters and
the exact response to invalid values may differ.