CTSKF-39 Allow main_hearing_date parameter
The main hearing date of a claim needs to be included in GET requests
made to Fee Calculator to using the /api/v1/fee-schemes/ endpoint to
allow the selection of CLAIR interim fee schemes.

This commit adds main_hearing_date as an optional parameter in the
SchemeListQuerySerializer class. It also clarifies the meaning of the
case date, adds extra tests, and updates the README.