Fix conflict between local/circleci test dependencies
It's not currently possible to run tests on a (clean) local machine. Tests
fail with the error "No module named xmlrunner".

This is because the `unittest-xml-reporting` wheel is only loaded in
`requirements/circleci.txt`. If a developer has not installed dependencies
from that file then tests fail locally. It is not necessarily obvious that
a developer should install CI dependencies for local work.

This commit moves the dependencies from `requirements/circleci.txt` into
a new `requirements/test.txt` file and includes that in both `circleci.txt`
and `local.txt`, which allows tests to run successfully locally and on
CircleCI.