Increase parellel test executors from 2 to 3
The test-calculator circleci job runs in parallel on two executors. As
we have added new fee schemes recently the number of tests has increased
and the tests now take > 12 minutes to complete.

This commit increases the number of executors to 3 which reduces the time
taken to run the calculator tests to about 8 minutes.