Use Pipenv in Dockefile and CircleCI pipeline
Replaces the use of pip and venv to build and test Fee Calculator with
Pipenv.

To do this in CircleCI a new executor has been added to .circleci/config.yml.
This is built on a cimg/python image and is used to run the linting,
testing and code coverage jobs that require Pipenv, to reduce the need
to install extra packages in the existing cloud-platform-executor.