Remove reviewers fields from dependabot.yml
Dependabot PRs are being annotated with the following comment:

```
The reviewers field in the dependabot.yml file will be removed soon.
Please use the code owners file to specify reviewers for Dependabot PRs
```

This commit removes the reviewers fields.

It also updates the CODEOWNERS file to remove the Claim for Payment
team as a reviewer - this team has not formally existed for some time
and all team members are also in the CLAIR Taskforce team.