Remove django-extended-choices
The django-extended-choices library imports pkg_resources, which has
been deprecated and is scheduled for removal as early as 30 November
2025.

This is causing issues upgrading to the latest version of Python.

The functionality offered by django-extended-choices can be replicated
using built in Django functionality, so this commit removes the library.
Subsequent commits will update the code.